Don't silently install Suggests:, fixes bug #47934
When installing a package which suggests an uninstalled package, urpmi
currently installs both packages without asking the user while it asks
when installing a package which requires an uninstalled package. This is
because we call resolve_requested__no_suggests to find which suggested
package to install, but this mark the package as requested as well (and
we don't ask for confirmation before installing a requested package).
This patch calls resolve_requested__no_suggests_ (with a trailing '_')
instead which is the same as resolve_requested__no_suggests but does not
mark the packages as requested.

git-svn-id: svn+ssh://svn.mandriva.com/svn/soft/rpm/perl-URPM/trunk@254572 99302b65-d5f7-0310-b3dd-f8cd6f4e3d94