keep psel/promote info and remove deadlocked pkg instead of aborting upgrade
Patch from Anssi Hannula, fixes bug #52105

lib64gcj9 has to be removed (conflicts), therefore so does libgcj9-src (x86_64)
that depends on it. Installed java-1.5.0-gcj-src depends on libgcj9-src, so
perl-URPM tries to promote the i586 libgcj9-src for it. However, strict_arch
check prohibits changing the arch, so it falls back to backtracking.
Backtracking finds libgcj9-src again and tries to select it; however, it drops
$dep->{psel} and $dep->{promote}, so when it fails again, perl-URPM does not
know java-1.5.0-gcj-src should also be removed due to the failed promotion.

This patch changes the code in backtrack_selected in two ways:

1) When a replacement package is found, keep {promote} and {psel} info, so that
when backtrack_selected is run the second time it correctly handles the failed
promotion.
2) When such a replacement fails as well and the deadlock protection is
triggered, do not switch to keep mode for the package that caused the promotion
(it would unselect all the packages involved in the update that caused the
promotion). Instead proceed to remove it as usual.

The patch introduces no failures on the urpmi testsuite.

git-svn-id: svn+ssh://svn.mandriva.com/svn/soft/rpm/perl-URPM/trunk@258634 99302b65-d5f7-0310-b3dd-f8cd6f4e3d94