unselect current package if an avoided package is already selected
Patch by Anssi Hannula, fixes bug #52145

If package 'a' Conflicts on 'b', and user (or a dependency chain) tries to
install both at the same time, perl-URPM will only detect the conflict if
package 'a' gets selected first (this depends on hdlist order), as
_set_rejected_from is only called in one direction from _handle_conflicts and
it does not detect that a package it is about to reject_from is already
selected.

This bug currently causes a failure in urpmi handle-conflict-deps2 testcase.

This commit checks if avoided package is already selected, and unselects
current package in such a case

git-svn-id: svn+ssh://svn.mandriva.com/svn/soft/rpm/perl-URPM/trunk@258640 99302b65-d5f7-0310-b3dd-f8cd6f4e3d94